+++ ContractManager/ContractBlueprint/Prerequisite.cs
@@ -19,6 +19,39 @@ namespace ContractManager.ContractBlueprint
         [XmlElement("maxNumAcceptedContracts", DataType = "unsignedInt")]
         public uint maxNumAcceptedContracts { get; set; } = uint.MaxValue;
 
+        // type: maxCompleteCount - offer contract if number of completed instances of this contract blueprint is less than this number.
+        [XmlElement("maxCompleteCount", DataType = "unsignedInt")]
+        public uint maxCompleteCount { get; set; } = 0;
+
+        // type: maxFailedCount - offer contract if number of completed instances of this contract blueprint is less than this number.
+        [XmlElement("maxFailedCount", DataType = "unsignedInt")]
+        public uint maxFailedCount { get; set; } = uint.MaxValue;
+
+        // type: maxConcurrentCount - offer contract if number of completed instances of this contract blueprint is less than this number.
+        [XmlElement("maxConcurrentCount", DataType = "unsignedInt")]
+        public uint maxConcurrentCount { get; set; } = 0;
+
+        // type: hasCompletedContract - offer contract if the specified contract blueprint uid has been completed.
+        [XmlElement("hasCompletedContract", DataType = "string")]
+        public string hasCompletedContract { get; set; }
+
+        // type: hasFailedContract - offer contract if the specified contract blueprint uid has been failed.
+        [XmlElement("hasFailedContract", DataType = "string")]
+        public string hasFailedContract { get; set; }
+
+        // type: hasAcceptedContract - offer contract if the specified contract blueprint uid has been accepted (and not yet completed/failed).
+        [XmlElement("hasAcceptedContract", DataType = "string")]
+        public string hasAcceptedContract { get; set; }
+
+        // type: minNumberOfVessels - offer contract if number of vessels in current celestial system is more than this number.
+        [XmlElement("minNumberOfVessels", DataType = "unsignedInt")]
+        public uint minNumberOfVessels { get; set; } = 0;
+
+        // type: maxNumberOfVessels - offer contract if number of vessels in current celestial system is less than this number.
+        [XmlElement("maxNumberOfVessels", DataType = "unsignedInt")]
+        public uint maxNumberOfVessels { get; set; } = uint.MaxValue;
+
+
         public Prerequisite() { }
         
         public void WriteToConsole()
@@ -36,6 +69,21 @@ namespace ContractManager.ContractBlueprint
         internal bool Validate()
         {
             // nothing to validate.
+            if (this.type == PrerequisiteType.HasCompletedContract && String.IsNullOrEmpty(this.hasCompletedContract))
+            {
+                Console.WriteLine("[CM] [WARNING] prerequisite with type 'hasCompletedContract' requires a contract blueprint uid to be defined in hasCompletedContract.");
+                return false;
+            }
+            if (this.type == PrerequisiteType.HasFailedContract && String.IsNullOrEmpty(this.hasFailedContract))
+            {
+                Console.WriteLine("[CM] [WARNING] prerequisite with type 'hasFailedContract' requires a contract blueprint uid to be defined in hasFailedContract.");
+                return false;
+            }
+            if (this.type == PrerequisiteType.HasAcceptedContract && String.IsNullOrEmpty(this.hasAcceptedContract))
+            {
+                Console.WriteLine("[CM] [WARNING] prerequisite with type 'hasAcceptedContract' requires a contract blueprint uid to be defined in hasAcceptedContract.");
+                return false;
+            }
             // PrerequisiteType doesn't need to be validated, loading XML will throw an exception.
             return true;
         }
@@ -46,6 +94,22 @@ namespace ContractManager.ContractBlueprint
         [XmlEnum("maxNumOfferedContracts")]
         MaxNumOfferedContracts,
         [XmlEnum("maxNumAcceptedContracts")]
-        MaxNumAcceptedContracts
+        MaxNumAcceptedContracts,
+        [XmlEnum("maxCompleteCount")]
+        MaxCompleteCount,
+        [XmlEnum("maxFailedCount")]
+        MaxFailedCount,
+        [XmlEnum("maxConcurrentCount")]
+        MaxConcurrentCount,
+        [XmlEnum("hasCompletedContract")]
+        HasCompletedContract,
+        [XmlEnum("hasFailedContract")]
+        HasFailedContract,
+        [XmlEnum("hasAcceptedContract")]
+        HasAcceptedContract,
+        [XmlEnum("minNumberOfVessels")]
+        MinNumberOfVessels,
+        [XmlEnum("maxNumberOfVessels")]
+        MaxNumberOfVessels
     }
 }